Updated file information utility class
Updated File Information utility class to provide getters and setters for modifying/retrieving file properties.

Added method to get the file size in a human readable format. If the size of the file is unknown, the method throws a NumberFormatException(the UI components that need this information will be provided with a string like "Unknown" in this case
to let the user know that the file size could not be retrieved by the application).