Added string constants for intent extras used in the file sharing process.Set default night mode for the app to MODE_NIGHT_YES(the application will use dark mode by default).
Added string constants for intent extras used in the file sharing process. After the user selects a file from the file picker or shares it in the application,details about the file will be provided
in the FileSharingActivity

If the user clicks the "Send" button, the required information is supplied as extras(uri,file name,size) for the intent that will start the FileConnection service.