Updated FileSharingActivity
Updated FileSharingActivity to get the information for the file and store it in a FileInformation object.
If the system cannot get the file name, a random UUID will be provided.
Iff the system cannot get the file size, the "Unknown" string will be provided.
Updated FileSharingActivity to use the two newly created fragments and switch between them when needed.
Updated FileSharingActivity to show a FloatingActionButton as the promoted action of the activity, replacing the old "Send File" button.
If no device is connected/the device disconnects, then the icon of the floating button will change to a laptop(letting the user know that he/she must connect to the PC first).
If the user click the button,the MainActivity will be opened.